pyproject.toml
[tool.poetry]
name = "value-blue-api"
version = "0.1.0"
description = ""
authors = ["VP"]
readme = "README.md"

packages = [
    { include = "models" },
    { include = "helpers" },
]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.1.1"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
behave = "^1.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
